+++ pmd/src/main/java/net/sourceforge/pmd/lang/java/symboltable/ClassScope.java
@@ -8,6 +8,9 @@ import java.util.List;
 import java.util.Map;
 
 import net.sourceforge.pmd.lang.ast.Node;
+import net.sourceforge.pmd.lang.java.ast.ASTArgumentList;
+import net.sourceforge.pmd.lang.java.ast.ASTFormalParameter;
+import net.sourceforge.pmd.lang.java.ast.ASTLiteral;
 import net.sourceforge.pmd.lang.java.ast.ASTName;
 import net.sourceforge.pmd.lang.symboltable.NameDeclaration;
 import net.sourceforge.pmd.lang.symboltable.NameOccurrence;
@@ -128,13 +131,28 @@ public class ClassScope extends AbstractJavaScope {
         if (occurrence.isMethodOrConstructorInvocation()) {
             for (MethodNameDeclaration mnd: methodDeclarations.keySet()) {
                 if (mnd.getImage().equals(occurrence.getImage())) {
-                    int args = occurrence.getArgumentCount();
-                    if (args == mnd.getParameterCount() || (mnd.isVarargs() && args >= mnd.getParameterCount() - 1)) {
-                        // FIXME if several methods have the same name
-                        // and parameter count, only one will get caught here
-                        // we need to make some attempt at type lookup and discrimination
-                        // or, failing that, mark this as a usage of all those methods
+                    List<String> parameterTypes = determineParameterTypes(mnd);
+                    List<String> argumentTypes = determineArgumentTypes(occurrence, parameterTypes);
+
+                    if (!mnd.isVarargs()
+                            && occurrence.getArgumentCount() == mnd.getParameterCount()
+                            && parameterTypes.equals(argumentTypes)) {
                         return mnd;
+                    } else if (mnd.isVarargs()) {
+                        int varArgIndex = parameterTypes.size() - 1;
+                        String varArgType = parameterTypes.get(varArgIndex);
+                        if (parameterTypes.subList(0, varArgIndex).equals(argumentTypes.subList(0, varArgIndex))) {
+                            boolean sameType = true;
+                            for (int i = varArgIndex; i < argumentTypes.size(); i++) {
+                                if (!varArgType.equals(argumentTypes.get(i))) {
+                                    sameType = false;
+                                    break;
+                                }
+                            }
+                            if (sameType) {
+                                return mnd;
+                            }
+                        }
                     }
                 }
             }
@@ -172,6 +190,84 @@ public class ClassScope extends AbstractJavaScope {
             }
         }
         return result;
+    }
+
+    /**
+     * Provide a list of types of the parameters of the given method declaration.
+     * The types are simple type images. 
+     * @param mnd the method declaration.
+     * @return List of types
+     */
+    private List<String> determineParameterTypes(MethodNameDeclaration mnd) {
+        List<String> parameterTypes = new ArrayList<String>();
+        List<ASTFormalParameter> parameters = mnd.getMethodNameDeclaratorNode().findDescendantsOfType(ASTFormalParameter.class);
+        for (ASTFormalParameter p : parameters) {
+            parameterTypes.add(p.getTypeNode().getTypeImage());
+        }
+        return parameterTypes;
+    }
+
+    /**
+     * Provide a list of types of the arguments of the given method call.
+     * The types are simple type images. If the argument type cannot be determined (e.g. because it is itself
+     * the result of a method call), the parameter type is used - so it is assumed, it is of the correct type.
+     * This might cause confusion when methods are overloaded.
+     * @param occurrence the method call
+     * @param parameterTypes the parameter types of the called method
+     * @return the list of argument types
+     */
+    private List<String> determineArgumentTypes(JavaNameOccurrence occurrence, List<String> parameterTypes) {
+        final String unknown_type = "unknown";
+        List<String> argumentTypes = new ArrayList<String>();
+        ASTArgumentList arguments = occurrence.getLocation().jjtGetParent().jjtGetParent().getFirstDescendantOfType(ASTArgumentList.class);
+        if (arguments != null) {
+            for (int i = 0; i < arguments.jjtGetNumChildren(); i++) {
+                ASTName name = arguments.jjtGetChild(i).getFirstDescendantOfType(ASTName.class);
+                String typeImage = unknown_type;
+                if (name != null) {
+                    Scope s = name.getScope();
+                    while (s != null) {
+                        if (s.contains(new JavaNameOccurrence(name, name.getImage()))) {
+                            break;
+                        }
+                        s = s.getParent();
+                    }
+                    if (s != null) {
+                        Map<VariableNameDeclaration, List<NameOccurrence>> vars = s.getDeclarations(VariableNameDeclaration.class);
+                        for (VariableNameDeclaration d : vars.keySet()) {
+                            if (d.getImage().equals(name.getImage())) {
+                                typeImage = d.getTypeImage();
+                                break;
+                            }
+                        }
+                    }
+                } else {
+                    ASTLiteral literal = arguments.jjtGetChild(i).getFirstDescendantOfType(ASTLiteral.class);
+                    if (literal != null) {
+                        if (literal.isCharLiteral()) {
+                            typeImage = "char";
+                        } else if (literal.isStringLiteral()) {
+                            typeImage = "String";
+                        } else if (literal.isFloatLiteral()) {
+                            typeImage = "float";
+                        } else if (literal.isIntLiteral()) {
+                            typeImage = "int";
+                        }
+                    }
+                }
+                argumentTypes.add(typeImage);
+            }
+        }
+        // replace the unknown type with the correct parameter type of the method.
+        // in case the argument is itself a method call, we can't determine the result type of the called
+        // method. Therefore the parameter type is used.
+        // This might cause confusion, if method overloading is used.
+        for (int i = 0; i < argumentTypes.size() && i < parameterTypes.size(); i++) {
+            if (unknown_type.equals(argumentTypes.get(i))) {
+                argumentTypes.set(i, parameterTypes.get(i));
+            }
+        }
+        return argumentTypes;
     }
 
     public String toString() {